# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAMAZIRT API Documentation",
        "description": "API for TAMAZIRT",
        "contact": {
            "email": "support@tamazirt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "TAMAZIRT API Server"
        }
    ],
    "paths": {
...
        "/api/navigator-profiles": {
            "get": {
                "tags": [
                    "Navigator Profiles"
                ],
                "summary": "Get all navigator profiles (Admin only)",
                "description": "Retrieves a list of all existing navigator profiles, including associated user data. Requires 'admin' role.",
                "operationId": "adminIndexProfiles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Fez"
                                                    },
                                                    "verified": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "user": {
                                                        "description": "Associated user details",
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Navigator Profiles"
                ],
                "summary": "Create a new navigator user and profile (Admin only)",
                "description": "Creates a User account with the 'navigator' role and an associated NavigatorProfile. Requires 'admin' role.",
                "operationId": "adminStoreNavigator",
                "requestBody": {
                    "description": "Data for the new navigator account and profile",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "city"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Hamid Navigator"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hamid.nav@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepass123"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Rabat"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Experienced local guide.",
                                        "nullable": true
                                    },
                                    "languages": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Arabic",
                                            "French"
                                        ],
                                        "nullable": true
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+212610203040",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Navigator account and profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Navigator account and profile created successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/navigators/{id}": {
            "put": {
                "tags": [
                    "Navigator Profiles"
                ],
                "summary": "Admin update for any navigator profile",
                "description": "Allows an administrator to update any navigator's user and profile details, including the 'verified' status.",
                "operationId": "adminUpdateNavigatorProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the navigator profile to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Updated Admin Name",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated.admin@example.com",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret123",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Casablanca",
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Updated bio by admin.",
                                        "nullable": true
                                    },
                                    "languages": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "nullable": true
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "verified": {
                                        "description": "Admin can change verification status.",
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully by admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Navigator profile updated successfully by admin."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/navigator-profile": {
            "get": {
                "tags": [
                    "Navigator Profiles"
                ],
                "summary": "Get own profile (Navigator only)",
                "description": "Retrieves the profile and user data associated with the authenticated navigator.",
                "operationId": "navigatorMyProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The navigator profile object with nested user details.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: navigator access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}